backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==5.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.9.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
marshmallow==3.22.0
